Extracted platform stuff to its own library
