+++ src/util/CMakeLists.txt
@@ -1,5 +1,10 @@
 set(solvespace_util_SOURCES
-        ss_util.h
-        ss_util.cpp ss_assert.cpp ss_assert.h)
-add_library(util ${solvespace_util_SOURCES})
+        ss_util.cpp)
+set(solvespace_util_HEADERS
+    ss_util.h
+    )
+add_library(util SHARED
+    ${solvespace_util_SOURCES}
+    ${solvespace_util_HEADERS}
+    )
 target_include_directories(util PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
