--- src/util/ss_assert.cpp
@@ -1,7 +1,0 @@
-//
-// Created by benjamin on 9/25/18.
-//
-
-#include "ss_assert.h"
-
-
