--- src/util/ss_assert.h
@@ -1,10 +1,0 @@
-//
-// Created by benjamin on 9/25/18.
-//
-
-#ifndef SOLVESPACE_SS_ASSERT_H
-#define SOLVESPACE_SS_ASSERT_H
-
-
-
-#endif //SOLVESPACE_SS_ASSERT_H
