+++ src/util/ss_util.h
@@ -14,6 +14,25 @@
 
 // defines
 
+// Debugging functions
+#if defined(__GNUC__)
+#define ssassert(condition, message) \
+    do { \
+        if(__builtin_expect((condition), true) == false) { \
+            SolveSpace::AssertFailure(__FILE__, __LINE__, __func__, #condition, message); \
+            __builtin_unreachable(); \
+        } \
+    } while(0)
+#else
+#define ssassert(condition, message) \
+    do { \
+        if((condition) == false) { \
+            SolveSpace::AssertFailure(__FILE__, __LINE__, __func__, #condition, message); \
+            abort(); \
+        } \
+    } while(0)
+#endif
+
 #ifndef isnan
 #   define isnan(x) (((x) != (x)) || (x > 1e11) || (x < -1e11))
 #endif
@@ -23,7 +42,11 @@
 namespace SolveSpace {
 
 // functions
-int64_t GetMilliseconds();
+#if defined(__GNUC__)
+__attribute__((noreturn))
+#endif
+void AssertFailure(const char *file, unsigned line, const char *function,
+                   const char *condition, const char *message);
 
 #if defined(__GNUC__)
 __attribute__((__format__ (__printf__, 1, 2)))
@@ -31,33 +54,5 @@ __attribute__((__format__ (__printf__, 1, 2)))
 std::string ssprintf(const char *fmt, ...);
 
 }
-
-//// Debugging functions
-//#if defined(__GNUC__)
-//#define ssassert(condition, message) \
-//    do { \
-//        if(__builtin_expect((condition), true) == false) { \
-//            SolveSpace::AssertFailure(__FILE__, __LINE__, __func__, #condition, message); \
-//            __builtin_unreachable(); \
-//        } \
-//    } while(0)
-//#else
-//#define ssassert(condition, message) \
-//    do { \
-//        if((condition) == false) { \
-//            SolveSpace::AssertFailure(__FILE__, __LINE__, __func__, #condition, message); \
-//            abort(); \
-//        } \
-//    } while(0)
-//#endif
-//
-//namespace SolveSpace {
-//
-//#if defined(__GNUC__)
-//__attribute__((noreturn))
-//#endif
-//void AssertFailure(const char *file, unsigned line, const char *function,
-//                   const char *condition, const char *message);
-//}
 
 #endif //SOLVESPACE_UTIL_H
